move safewriter pkg
