+++ src/go/pkg/safewriter/writer.go
