+++ src/go/plugin/go.d/agent/agent.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/netdata/netdata/go/plugins/logger"
 	"github.com/netdata/netdata/go/plugins/pkg/netdataapi"
+	"github.com/netdata/netdata/go/plugins/pkg/safewriter"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/confgroup"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/filelock"
@@ -21,7 +22,6 @@ import (
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/functions"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/jobmgr"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/module"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/safewriter"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/pkg/multipath"
 
 	"github.com/mattn/go-isatty"
