+++ src/go/plugin/go.d/agent/functions/manager.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/netdata/netdata/go/plugins/logger"
 	"github.com/netdata/netdata/go/plugins/pkg/netdataapi"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/safewriter"
+	"github.com/netdata/netdata/go/plugins/pkg/safewriter"
 )
 
 func NewManager() *Manager {
