+++ src/go/plugin/go.d/agent/jobmgr/manager.go
@@ -13,10 +13,10 @@ import (
 
 	"github.com/netdata/netdata/go/plugins/logger"
 	"github.com/netdata/netdata/go/plugins/pkg/netdataapi"
+	"github.com/netdata/netdata/go/plugins/pkg/safewriter"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/confgroup"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/functions"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/module"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/safewriter"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/ticker"
 
 	"github.com/mattn/go-isatty"
