+++ src/go/plugin/go.d/agent/jobmgr/sim_test.go
@@ -11,9 +11,9 @@ import (
 	"time"
 
 	"github.com/netdata/netdata/go/plugins/pkg/netdataapi"
+	"github.com/netdata/netdata/go/plugins/pkg/safewriter"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/confgroup"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/module"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/safewriter"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
